Misc updates
- added examples section. moved connection.cpp to the examples subproject.
- minor tweaks to tests.
- Adding warnings about python3 and so on if you try to run the full test suite.
- upstream tests still not compiling due to code generation issues.
- renamed rethinkdb.hpp => rethinkdb.h

+++ test/test.cc
@@ -43,7 +43,7 @@ void test_cursor() {
     R::Cursor cursor = R::range(10000).run(*conn);
     TEST_EQ(cursor.next(), 0);
     R::Array array = cursor.to_array();
-    TEST_EQ(array.size(), 9999);
+    TEST_EQ(static_cast<uint64_t>(array.size()), 9999);
     TEST_EQ(*array.begin(), 1);
     TEST_EQ(*array.rbegin(), 9999);
     int i = 0;
